Feat: add 2024 contributions (untested)
There are several contributions in here that were not tracked. Since the
project is still in the initial development stage, work will be done to
split this commit into parts.

+++ lib/libc/string/strncmp.c
@@ -9,12 +9,14 @@
  * Copyright (c) 2024 George Witt
  * SPDX-License-Identifier: NOASSERTION
  */
+#include <stdint.h>
+#include <stddef.h>
 #include <string.h>
 
 int strncmp(const char *s1, const char *s2, size_t n)
 {
-   for (size_t i = 0; i < n; i++) {
-      int cmp = (s1 + i) - (s2 + i);
+   for (ptrdiff_t i = 0; i < n; i++) {
+      ptrdiff_t cmp = (s1 + i) - (s2 + i);
       if (cmp != 0) {
          return cmp;
       }
